refactor: rename division to integer_division to be in line with multiplication
