+++ pumpkin-crates/core/src/propagators/arithmetic/mod.rs
@@ -1,6 +1,6 @@
 pub(crate) mod absolute_value;
 pub(crate) mod binary;
-pub(crate) mod division;
+pub(crate) mod integer_division;
 pub(crate) mod integer_multiplication;
 pub(crate) mod linear_less_or_equal;
 pub(crate) mod linear_not_equal;
